Updated class names to MultipleExchange* and updated version number
