--- Source/EasyNetQ.Tests/AdvancedMessagePolymorphismTest/MessageWithTwoInterfaces.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace EasyNetQ.Tests.AdvancedMessagePolymorphismTest
-{
-    public class MessageWithTwoInterfaces : IMessageInterfaceOne, IMessageInterfaceTwo
-    {
-    }
-}
