+++ Source/EasyNetQ.Tests/EasyNetQ.Tests.csproj
@@ -89,10 +89,10 @@
       <Link>Properties\Version.cs</Link>
     </Compile>
     <Compile Include="AdvancedBusEventHandlersTests.cs" />
-    <Compile Include="AdvancedMessagePolymorphismTest\AdvancedPolymorphismPublishExchangeDeclareStrategyTests.cs" />
-    <Compile Include="AdvancedMessagePolymorphismTest\IMessageInterfaceTwo.cs" />
-    <Compile Include="AdvancedMessagePolymorphismTest\IMessageInterfaceOne.cs" />
-    <Compile Include="AdvancedMessagePolymorphismTest\MessageWithTwoInterfaces.cs" />
+    <Compile Include="MultipleExchangeTest\MultipleExchangePublishExchangeDeclareStrategyTests.cs" />
+    <Compile Include="MultipleExchangeTest\IMessageInterfaceTwo.cs" />
+    <Compile Include="MultipleExchangeTest\IMessageInterfaceOne.cs" />
+    <Compile Include="MultipleExchangeTest\MessageWithTwoInterfaces.cs" />
     <Compile Include="AmqpExceptionParserTests.cs" />
     <Compile Include="AutoSubscriberTests\When_autosubscribing_with_assembly_scanning.cs" />
     <Compile Include="AutoSubscriberTests\When_autosubscribing_with_subscription_configuration_action_and_attribute.cs" />
