+++ Source/EasyNetQ.Tests/MultipleExchangeTest/IMessageInterfaceOne.cs
@@ -1,0 +1,6 @@
+﻿namespace EasyNetQ.Tests.MultipleExchangeTest
+{
+    public interface IMessageInterfaceOne
+    {
+    }
+}
