+++ Source/EasyNetQ.Tests/MultipleExchangeTest/IMessageInterfaceTwo.cs
@@ -1,0 +1,6 @@
+﻿namespace EasyNetQ.Tests.MultipleExchangeTest
+{
+    public interface IMessageInterfaceTwo
+    {
+    }
+}
