+++ Source/EasyNetQ.Tests/MultipleExchangeTest/MessageWithTwoInterfaces.cs
@@ -1,0 +1,6 @@
+﻿namespace EasyNetQ.Tests.MultipleExchangeTest
+{
+    public class MessageWithTwoInterfaces : IMessageInterfaceOne, IMessageInterfaceTwo
+    {
+    }
+}
