--- Source/EasyNetQ/AdvancedMessagePolymorphism/AdvancedMessagePolymorphismExtension.cs
@@ -1,13 +1,0 @@
-﻿using EasyNetQ.Producer;
-
-namespace EasyNetQ.AdvancedMessagePolymorphism
-{
-    public static class AdvancedMessagePolymorphismExtension
-    {
-        public static IServiceRegister EnableAdvancedMessagePolymorphism(this IServiceRegister serviceRegister)
-        {
-            return serviceRegister
-                .Register<IPublishExchangeDeclareStrategy, AdvancedPolymorphismPublishExchangeDeclareStrategy>();
-        }
-    }
-}
